LPS-66657 Move upgrade
