+++ modules/apps/forms-and-workflow/dynamic-data-lists/dynamic-data-lists-service/src/main/java/com/liferay/dynamic/data/lists/internal/upgrade/DDLServiceUpgrade.java
@@ -16,8 +16,8 @@ package com.liferay.dynamic.data.lists.internal.upgrade;
 
 import com.liferay.dynamic.data.lists.internal.upgrade.v1_0_0.UpgradeKernelPackage;
 import com.liferay.dynamic.data.lists.internal.upgrade.v1_0_0.UpgradeLastPublishDate;
-import com.liferay.dynamic.data.lists.internal.upgrade.v1_0_0.UpgradeRecordGroupId;
 import com.liferay.dynamic.data.lists.internal.upgrade.v1_0_0.UpgradeSchema;
+import com.liferay.dynamic.data.lists.internal.upgrade.v1_0_1.UpgradeRecordGroupId;
 import com.liferay.dynamic.data.mapping.service.DDMStructureLocalService;
 import com.liferay.portal.upgrade.registry.UpgradeStepRegistrator;
